requirements.txt
pyserial
requests
tk